+++ client/export_test.go
@@ -15,9 +15,12 @@
 package client
 
 import (
+	"bytes"
+	"encoding/json"
 	"fmt"
 	"io"
 	"net/url"
+	"time"
 )
 
 var (
@@ -45,6 +48,35 @@ func (client *Client) SetGetWebsocket(f getWebsocketFunc) {
 	client.getWebsocket = f
 }
 
+// DebugGet sends a GET debug action to the server with the provided parameters.
+func (client *Client) DebugGet(action string, result interface{}, params map[string]string) error {
+	urlParams := url.Values{"action": []string{action}}
+	for k, v := range params {
+		urlParams.Set(k, v)
+	}
+	_, err := client.doSync("GET", "/v1/debug", urlParams, nil, nil, &result)
+	return err
+}
+
+type debugAction struct {
+	Action string      `json:"action"`
+	Params interface{} `json:"params,omitempty"`
+}
+
+// DebugPost sends a POST debug action to the server with the provided parameters.
+func (client *Client) DebugPost(action string, params interface{}, result interface{}) error {
+	body, err := json.Marshal(debugAction{
+		Action: action,
+		Params: params,
+	})
+	if err != nil {
+		return err
+	}
+
+	_, err = client.doSync("POST", "/v1/debug", nil, nil, bytes.NewReader(body), result)
+	return err
+}
+
 // WaitStdinDone waits for WebsocketSendStream to be finished calling
 // WriteMessage to avoid a race condition.
 func (p *ExecProcess) WaitStdinDone() {
@@ -52,3 +84,15 @@ func (p *ExecProcess) WaitStdinDone() {
 }
 
 type ClientWebsocket = clientWebsocket
+
+// FakeDoRetry fakes the delays used by the do retry loop.
+func FakeDoRetry(retry, timeout time.Duration) (restore func()) {
+	oldRetry := doRetry
+	oldTimeout := doTimeout
+	doRetry = retry
+	doTimeout = timeout
+	return func() {
+		doRetry = oldRetry
+		doTimeout = oldTimeout
+	}
+}
